refactor(cli): extract start/stop/restart daemon logic to commands/start.rs
Move daemonize, stop_process, and start_bot from main.rs into a dedicated
commands/start module with pub(crate) entry points run_start, run_up,
run_stop, and run_restart. Add // SAFETY: comments to both unsafe libc::kill
blocks. Remove now-unused imports from main.rs.
